Painting my house
This is code to create a template for painting the walls of my house.
The morphospace of landmarks from 3,319 Passiflora leaves (see
https://github.com/DanChitwood/PassifloraLeaves for data) was projected
onto a Fibonacci phyllotactic spiral (made with code from here
http://stackoverflow.com/questions/26141237/how-to-create-phyllotaxis-sp
irals-with-r citing Algorithmic Botany
http://algorithmicbotany.org/papers/abop/abop-ch4.pdf). The ultimate
goal was to order the leaves by their first principal component values
from the first level of the house to the second so that walking up the
stairs the morphospace could be experienced. The arrangement by angle
and position using phyllotactic patterning is to evoke a sensation of
leaves growing from a shoot tip. The coded template will be light
projected onto the wall and then hand painted.